Fixed refactoring bugs
