+++ src/main.js
@@ -38,7 +38,7 @@ function createWindow() {
     });
 
     win.webContents.setVisualZoomLevelLimits(1, 5);
-    Menu.setApplicationMenu(createMenuBar(config));
+    Menu.setApplicationMenu(createMenuBar(config, win));
 
     // and load the index.html of the app.
     win.loadURL(
@@ -79,11 +79,27 @@ function createWindow() {
     mainWindowState.manage(win);
 }
 
-util.instanceLock();
+var gotTheLock = app.requestSingleInstanceLock();
+
+if (!gotTheLock) {
+    app.quit();
+} else {
+    app.on("second-instance", () => {
+        // Someone tried to run a second instance, we should focus our window.
+        if (win) {
+            if (win.isMinimized()) {
+                win.restore();
+                win.focus();
+            }
+            win.show();
+            win.focus();
+        }
+    });
+}
 
 app.on("ready", () => {
     createWindow();
-    createTray(config);
+    createTray(config, win);
 
     win.webContents.on("dom-ready", () => {
         if (config["beta"]) {
